update/cleanup strings.po
On strings.po for en_gb was on some places 'msgstr ""' set where should be on
this language empty. Further was the related text there different to other and
more detailed, for this reason is this text now moved to new "msgctxt" and to
use as help text on addon settings.